+++ CHANGELOG.md
@@ -281,6 +281,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Query format constraints
 - Feasibility service and endpoint for Lens-style AST_DATA queries through Beam and Focus
 - ENABLE_FEASIBILITY feature toggle
+- Form title canonical order
 
 
 ### Changed